Remove the old Editor in favor of State
