+++ src/editor/state/ui.rs
@@ -3,7 +3,7 @@ mod side_panel;
 mod toolbar;
 mod windows;
 
-impl super::State {
+impl super::Editor {
     pub fn ui(&mut self, egui_ctx: &egui::Context) {
         self.draw_toolbar(egui_ctx);
         self.draw_side_panel(egui_ctx);
